+++ pkg/server/render_test.go
@@ -1,8 +1,6 @@
 package server
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
 	"io"
 	"net/http"
@@ -26,11 +24,13 @@ import (
 
 var _ = Describe("server", func() {
 	var httpServer *httptest.Server
+	var s *storage.Storage
 
 	testing.WithConfig(func(cfg **config.Config) {
 		BeforeEach(func() {
 			(*cfg).Server.APIBindAddr = ":10044"
-			s, err := storage.New(storage.NewConfig(&(*cfg).Server), logrus.StandardLogger(), prometheus.NewRegistry(), new(health.Controller))
+			var err error
+			s, err = storage.New(storage.NewConfig(&(*cfg).Server), logrus.StandardLogger(), prometheus.NewRegistry(), new(health.Controller))
 			Expect(err).ToNot(HaveOccurred())
 			e, _ := exporter.NewExporter(nil, nil)
 			c, _ := New(Config{
@@ -45,6 +45,10 @@ var _ = Describe("server", func() {
 			})
 			h, _ := c.serverMux()
 			httpServer = httptest.NewServer(h)
+		})
+		JustAfterEach(func() {
+			s.Close()
+			httpServer.Close()
 		})
 		Context("/render", func() {
 			It("supports name and query parameters", func() {
@@ -93,9 +97,3 @@ var _ = Describe("server", func() {
 		})
 	})
 })
-
-func reqBody(v interface{}) io.Reader {
-	var b bytes.Buffer
-	Expect(json.NewEncoder(&b).Encode(v)).ToNot(HaveOccurred())
-	return &b
-}