+++ packages/pyroscope-flamegraph/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.18.5](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.18.4...@pyroscope/flamegraph@0.18.5) (2022-08-11)
+
+
+### Bug Fixes
+
+* Make trace units be time based ([#1387](https://github.com/pyroscope-io/pyroscope/issues/1387)) ([a567c2c](https://github.com/pyroscope-io/pyroscope/commit/a567c2c1877c462a6a4dceda87455017c1cf7276))
+
+
+
+
+
 ## [0.18.4](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.18.3...@pyroscope/flamegraph@0.18.4) (2022-08-11)
 
 **Note:** Version bump only for package @pyroscope/flamegraph